USBDevice: do not use automatic multi-packet USB transactions
If the host sends an unfinished OUT transaction (=> a transaction
that ends with a data packet filled to the maximum allowed size but
not followed by another shorter packet or a ZLP) the Transaction
Completed interrupt is not fired and the incoming data remains
suspendend until the host sends more data.

This commit fix the problem by disabilng multi-packet transactions.